+++ addOns/requester/CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Added
+- Persist tabs (with name, order and message) in session database
 
 ## [7.1.0] - 2022-12-19
 ### Added